extracted updateCardInventoryWithFlag
