+++ monolith/routes/auth.ts
@@ -19,6 +19,7 @@ import getFormatLegalities from '../interactors/getFormatLegalities';
 import addCardToInventoryReceiving from '../interactors/addCardToInventoryReceiving';
 import getSuspendedSales from '../interactors/getSuspendedSales';
 import getSuspendedSale from '../interactors/getSuspendedSale';
+import updateCardInventoryWithFlag from '../interactors/updateCardInventoryWithFlag';
 
 interface RequestWithUserInfo extends Request {
     locations: string[];
@@ -347,64 +348,6 @@ async function validateInventory(
         }
 
         return true;
-    } catch (e) {
-        throw e;
-    } finally {
-        await client.close();
-    }
-}
-
-/**
- * Updates card inventory based on the card's passed properties (qtyToSell, finishCondition, id, name) and CHANGE_FLAG
- * @param {Object} card - the card involved in the transaction
- * @param {String} CHANGE_FLAG = `INC` or `DEC`, determines whether to increase or decrease quantity, used for reserving inventory in suspension
- */
-async function updateCardInventoryWithFlag(
-    card,
-    CHANGE_FLAG,
-    location: ClubhouseLocation
-) {
-    const { qtyToSell, finishCondition, id, name } = card;
-    let quantityChange;
-
-    if (CHANGE_FLAG === 'DEC') {
-        quantityChange = -Math.abs(qtyToSell);
-    } else if (CHANGE_FLAG === 'INC') {
-        quantityChange = Math.abs(qtyToSell);
-    } else {
-        throw new Error('CHANGE_FLAG was not provided');
-    }
-
-    const client = await new MongoClient(process.env.MONGO_URI, mongoOptions);
-
-    try {
-        await client.connect();
-
-        console.log(
-            `Suspend sale, ${CHANGE_FLAG}: QTY: ${qtyToSell}, ${finishCondition}, ${name}, ${id}`
-        );
-
-        const db = client
-            .db(DATABASE_NAME)
-            .collection(collectionFromLocation(location).cardInventory);
-
-        await db.updateOne(
-            { _id: id },
-            {
-                $inc: {
-                    [`qoh.${finishCondition}`]: quantityChange,
-                },
-            }
-        );
-
-        // Validate inventory quantites to never be negative numbers
-        return await db.updateOne(
-            {
-                _id: id,
-                [`qoh.${finishCondition}`]: { $lt: 0 },
-            },
-            { $set: { [`qoh.${finishCondition}`]: 0 } }
-        );
     } catch (e) {
         throw e;
     } finally {
